requirements.txt
pandas 
# notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
ensure==1.0.2
joblib
types-PyYAML
fastapi
uvicorn
requests
mlflow
dagshub
ipykernel
seaborn
dvc
fastapi
uvicorn
streamlit
-e.